refactor(tests): use js playwright and fix eslint errors
