--- playwright.config.ts
@@ -1,19 +1,0 @@
-import {devices} from '@playwright/test';
-import type {PlaywrightTestConfig} from '@playwright/test';
-
-const config: PlaywrightTestConfig = {
-  webServer: {
-    port: 3000,
-    command: 'yarn docusaurus serve',
-  },
-  projects: [
-    {
-      name: 'chromium',
-      use: {
-        ...devices['Desktop Chrome'],
-      },
-    },
-  ],
-};
-
-export default config;
