+++ tests/screenshot.css
@@ -10,12 +10,14 @@ img[src$='.gif'],
 [class*='playgroundPreview'] {
   visibility: hidden;
 }
+
 /* Different docs last-update dates can alter layout */
 .theme-last-updated,
 /* Mermaid diagrams are rendered client-side and produce layout shifts */
 .docusaurus-mermaid-container {
   display: none;
 }
+
 /* Videos are laggy and the loading spinner is shown as change on Argos */
 video {
   visibility: hidden;
